backend/requirements.txt
httplib2==0.22.0
cachetools==5.5.0
pip==24.2
oauthlib==3.2.2
requests-oauthlib==2.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
requests==2.32.3
SQLAlchemy==2.0.36
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
google==3.0.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
python-dotenv==1.0.1